Print a timestamp as a comment every ~5 seconds
Our timestamps are relative to the start time but every once in a while
it's useful to know the local time for a specific event. In particular
for long-running recordings it allows a user to check the wall clock and
then go back to the recording to narrow down when something happened.