hid-replay: handle GetFeature/SetFeature requests from the kernel
This fixes the slow-to-start issue seen before, caused by the kernel
sending us GetFeature/SetFeature requests and eventually timing out.

Fix this by replying with errors to either request which shortcuts the
timeouts - good enough or hid-replay though.

uhid-virt doesn't give us access to the internally used file descriptor
so we can't poll it - instead we read an event and skip over the error,
if any.

To know when we're ready and no longer have to reply to events we check
if the expected sysfs path exists and the hidraw directory within. Since
we can't predict the last part of the path use a glob and check all of
them.