Print a color marker based on the Report ID
In the report summary and the input reports as read from the device
print a color marker so it's easy to associate two reports to each
other.

There's a static set of 7 colors and we simply modulo the report ID.
That's good enough for most devices with a small set of report ID and
using 7 means we get different colors for e.g. logitech devices that use
report IDs of 16 and 32.