build.rs: don't re-implement which, try to compile instead
Instead of reimplementing which let's just try to compile things and
fall back to our BPFs if the compilation fails. This is more robust than
trying re-create whatever libbpf_cargo does in its clang invocation.

Unfortunately the error itself is just a anyhow::bail so we can't detect
a specific "compiler is missing" error unless we want to dig around in
the error messages.