libinput: add support for reading hidraw events from a recording
Slightly more difficult because libinput record theoretically supports
having multiple hidraw nodes per device so the format is:

events:
  # An evdev event
  - evdev:
    - [  1, 192005,   2,   0,       8]
    - [  1, 192005,   0,   0,       0]
  # A HID report from hidraw1
  - hid:
      time: [  1, 207144]
      hidraw1: [
        0x1a, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
      ]

So we need to filter for hid, then within that filter for the first
hidraw node. This fails currently if we ever alternate hidraw nodes but
we're already very far into niche territory, so meh.