Handle UsagePages during Push/Pop
This only affects the comment in initial listing of the HID items in the
report descriptor where the last seen UsagePage was used for any
subsequent UsageId even after a Pop.

   UsagePage A       // UsagePage A
   Push
      UsagePage B    // UsagePage B
      ...
   Pop
   UsageId U         // Usage B(U)     <- should be A(U)

The UsagePage B should've been dropped with the Pop.

Fix this by adding push/pop handling as a proper stack.

Closes #46